requirements.txt
flask==2.0.1
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
flask-wtf==0.15.1 